requirements.txt
flet
opencv-python
python-dotenv
mysql-connector-python
bcrypt
pytz
jdatetime